requirements.txt
Pillow
audiotsm
scipy
numpy
pytube
moviepy
whisper
aiohttp
